Adds filter to $post_status
Currently the post status is limited to either publish or pending, depending on whether the sender's email address is linked to an existing account.

An extending plugin might want to hook in to 'publish_phone' to do stuff to the post before publishing, eg. add custom fields based on the message title or content.